pyproject.toml
[project]

name = "Mitzuki"
version = "0.1.0"
description = "Mizuki es tu asistente lunar perfecta para Discord. Con su elegancia celestial y tecnología avanzada, te ayuda a gestionar tu servidor con la suavidad de la luz de luna."
authors = [
    { name = "ItsJhonAlex", email = "itsjhonalex@gmail.com"}
]
license = { text = "MIT" }
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.2",
    "aiosqlite>=0.21.0",
    "colorama>=0.4.6",
    "discord-py==2.6.4",
    "pydantic>=2.12.4",
    "pynacl>=1.6.1",
    "python-dotenv>=1.2.1",
]

[project.scripts]
mizuki-bot = "main:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["bot*"]
exclude = ["*tests*", "*scripts*"]
